Added UI to start and restart the game and to keep track of lives and score

+++ Unity Project/COMP140 Game/Assets/Scripts/TrainController.cs
@@ -9,7 +9,7 @@ public class TrainController : MonoBehaviour
     private GameObject controller;
 
 
-    public void initialisation(int newSpeed, GameObject newController)
+    public void initialisation(float newSpeed, GameObject newController)
     {
         speed = newSpeed;
         controller = newController;
